backend/requirements.txt
celery==5.3.6
Flask==2.2.5
Flask_Cors==4.0.0
Flask_JWT_Extended==4.6.0
flask_restx==1.3.0
Flask_Security==3.0.0
flask_sqlalchemy==3.1.1
Jinja2==3.1.3
matplotlib==3.8.4
pdfkit==1.0.0
Pillow==9.0.1
Pillow==10.3.0
pyshorteners==1.0.1
PyYAML==5.4.1
PyYAML==6.0.1
Requests==2.31.0
SQLAlchemy==2.0.28
Werkzeug==3.0.2
wordcloud==1.9.3
